# backend/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Dermatology EHR API",
    "version": "1.0.0",
    "description": "Comprehensive API documentation for the Dermatology EHR system",
    "contact": {
      "name": "API Support",
      "email": "support@example.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "API Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT authorization token"
      },
      "tenantHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "x-tenant-id",
        "description": "Tenant ID header for multi-tenancy"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "Error message"
          }
        }
      },
      "ValidationError": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "description": "Validation error details from Zod"
          }
        }
      },
...
      "Appointment": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "scheduledStart": {
            "type": "string",
            "format": "date-time"
          },
          "scheduledEnd": {
            "type": "string",
            "format": "date-time"
          },
          "status": {
            "type": "string",
            "enum": [
              "scheduled",
              "confirmed",
              "checked-in",
              "in-progress",
              "completed",
              "cancelled",
              "no-show"
            ]
          },
          "patientId": {
            "type": "string",
            "format": "uuid"
          },
          "providerId": {
            "type": "string",
            "format": "uuid"
          },
          "locationId": {
            "type": "string",
            "format": "uuid"
          },
          "appointmentTypeId": {
            "type": "string",
            "format": "uuid"
          },
          "patientName": {
            "type": "string"
          },
          "providerName": {
            "type": "string"
          },
          "locationName": {
            "type": "string"
          },
          "appointmentTypeName": {
            "type": "string"
          },
          "durationMinutes": {
            "type": "integer"
          }
        }
      },
      "CreateAppointmentRequest": {
        "type": "object",
        "required": [
          "patientId",
          "providerId",
          "locationId",
          "appointmentTypeId",
          "scheduledStart",
          "scheduledEnd"
        ],
        "properties": {
          "patientId": {
            "type": "string",
            "minLength": 1
          },
          "providerId": {
            "type": "string",
            "minLength": 1
          },
          "locationId": {
            "type": "string",
            "minLength": 1
          },
          "appointmentTypeId": {
            "type": "string",
            "minLength": 1
          },
          "scheduledStart": {
            "type": "string",
            "format": "date-time"
          },
          "scheduledEnd": {
            "type": "string",
            "format": "date-time"
          },
          "status": {
            "type": "string"
          }
        }
      },
...
    }
  },
  "security": [
    {
      "bearerAuth": [],
      "tenantHeader": []
    }
  ],
  "paths": {
    "/api/appointments": {
      "get": {
        "summary": "List appointments",
        "description": "Retrieve appointments with optional filters for patient, date, or date range (limited to 200).",
        "tags": [
          "Appointments"
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {
            "tenantHeader": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "patientId",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Filter by patient ID"
          },
          {
            "in": "query",
            "name": "date",
            "schema": {
              "type": "string",
              "format": "date"
            },
            "description": "Filter by specific date (YYYY-MM-DD)"
          },
          {
            "in": "query",
            "name": "startDate",
            "schema": {
              "type": "string",
              "format": "date"
            },
            "description": "Filter by start date (YYYY-MM-DD)"
          },
          {
            "in": "query",
            "name": "endDate",
            "schema": {
              "type": "string",
              "format": "date"
            },
            "description": "Filter by end date (YYYY-MM-DD)"
          }
        ],
        "responses": {
          "200": {
            "description": "List of appointments",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "appointments": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Appointment"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create an appointment",
        "description": "Create a new appointment. Checks for provider time conflicts. Requires admin, front_desk, ma, or provider role.",
        "tags": [
          "Appointments"
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {
            "tenantHeader": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAppointmentRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Appointment created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "format": "uuid"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationError"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - Insufficient permissions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Time conflict for provider",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}